Fix closed toc eating clicks with phantom div
Previous behaviour was to remove the toc from view by applying a transform to the .toc > ul but *not* .toc itself.

This meant that .toc was still covering the body when the toc was "closed", eating clicks on anything that would be covered by the toc when opened.

This change puts a transform on .toc itself (and changes the transforms on the ul to match), fixing the issue. However, this creates a new stacking context, which affects .toctogglelabel, since its fixed position is now with respect to .toc, not the viewport. Setting .toctogglelabel { right: -100vw } keeps it in the same place.