feat: remove overflow-x hidden from body,html (#267)
Removing this is necessary so that position sticky can be used on the page.

Reference: https://stackoverflow.com/a/47745433

This was presumably added because of the skin having some errant elements at
some point. The only case I can find of this is .catlinks and #p-namespaces
in an earlier patch, but their layout has since been changed.