feat: tweak toc
3 different concerns addressed:

First: stop y-scrollbar on collapsible .toc extending into the .toctitle.

This looks weird because the rounded corner exposes the hack; and it makes the
scrollbar not accurately match its apparent position, because a part of it is
hidden. Hacking around with the box shadow is less visible.

Second: make .toc always fill out available space; add padding-right to the li elems.

In some cases, more often on Firefox, there's plenty of space for the .toc,
but it's not used. Instead we get a super thin .toc with text rubbing right up
against the scrollbar. Given that the layout already allocates space for the
.toc (from --width-toc), no reason not to just always use it with `width:
100%`.

Third: remove color change to li.active children.

The border already sufficiently encompasses all the children, so it's
redundant. The problem with it is that the color change is the same as the
hover color, which is the only visual indicator of which toc item is hovered.