fix(menu): 🐛 avoid double view button on talk page
In 9969479, we forced the view button to be visible to allow the return action (#845).
There are no good ways to target revision page only so we rely on the diff styles.
However, Extension:DiscussionTools inline the said styles on talk page while talk page also has
a return action registered under the associated page menu. This leaves two button serving the
same function.